requirements.txt
streamlit
pymupdf
faiss-cpu
sentence-transformers
transformers
huggingface-hub
